+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Features
 
+- Add JSON column sub-paths to column selection in query builder
+- Added events support in trace detail view.(https://github.com/grafana/clickhouse-datasource/pull/1128)
+
+## 4.6.0
+
+### Features
+
 - Add support for new Variant, Dynamic, and JSON types (https://github.com/grafana/clickhouse-datasource/pull/1108)
 
 ### Fixes